fix(Swap): decoupled swap-receive switch
