+++ packages/blockchain-wallet-v4-frontend/src/scenes/Exchange/ExchangeForm/SwapReceiveSwitch/index.js
@@ -1,0 +1,1 @@
+export { default } from './SwapReceiveSwitch'
