+++ packages/blockchain-wallet-v4-frontend/src/scenes/Exchange/ExchangeForm/SwapReceiveSwitch/selectors.js
@@ -1,0 +1,13 @@
+import { createSelector } from 'reselect'
+
+import { model } from 'data'
+
+const { sourceActive, targetActive } = model.rates
+
+export const getData = createSelector(
+  (state, { fix }) => fix,
+  fix => ({
+    sourceActive: sourceActive(fix),
+    targetActive: targetActive(fix)
+  })
+)
