+++ packages/blockchain-wallet-v4-frontend/src/scenes/Exchange/ExchangeForm/VerificationInfo/VerificationInfo.spec.js
@@ -4,7 +4,6 @@ import { mount } from 'enzyme'
 
 import { actions } from 'data'
 import { TIERS_STATES } from 'data/modules/profile/model'
-import {} from 'data/components/identityVerification/model'
 import profileReducer from 'data/modules/profile/reducers'
 
 import LimitInfo from './index'
